# docs/openapi.yml
openapi: 3.1.0
info:
  title: Card Collection Price Tracker API
  description: 'REST API for tracking Magic: The Gathering card prices'
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 0.1.0
paths:
...
  /collection:
    get:
      tags:
      - collection
      operationId: get_collection
      parameters:
      - name: page
        in: query
        description: Page number, 0-based (default 0). `page * page_size` must not exceed 10000
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
      - name: page_size
        in: query
        description: Items per page, 1 to 100 (default 20)
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
      - name: sort_by
        in: query
        description: Sort field
        required: false
        schema:
          $ref: '#/components/schemas/SortByParam'
      - name: sort_dir
        in: query
        description: Sort direction
        required: false
        schema:
          $ref: '#/components/schemas/SortDirParam'
      - name: q
        in: query
        description: Fuzzy search on card name or set
        required: false
        schema:
          type: string
      - name: rarity
        in: query
        description: Rarity codes, repeated for multiple values (e.g. rarity=C&rarity=U)
        required: false
        schema:
          type: array
          items:
            $ref: '#/components/schemas/RarityCodeParam'
      - name: sets
        in: query
        description: Comma-separated set codes
        required: false
        schema:
          type: string
      - name: price_min
        in: query
        description: Minimum trend price in cents
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
      - name: price_max
        in: query
        description: Maximum trend price in cents
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
      responses:
        '200':
          description: Paginated card collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCollectionResponse'
        '400':
          description: Pagination out of bounds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_auth: []
  /collection/import:
    get:
      tags:
      - collection
      operationId: list_card_imports
      responses:
        '200':
          description: The caller's imports, most recent first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardImportResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_auth: []
    post:
      tags:
      - collection
      operationId: import_cards
      requestBody:
        description: ManaBox CSV content (max 10 MB)
        content:
          text/plain:
            schema:
              type: string
        required: true
      responses:
        '202':
          description: Import accepted, processing in the background
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardImportStartedResponse'
        '400':
          description: Rejected file; `code` is `empty_file`, `malformed_csv`, `binder_export` or `unrecognized_format`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: An import is already in progress for this user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_auth: []
  /collection/import/{id}:
    get:
      tags:
      - collection
      operationId: get_card_import
      parameters:
      - name: id
        in: path
        description: Import id
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Import status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardImportResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Import not found, or not owned by the caller
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_auth: []
  /collection/price-history:
    get:
      tags:
      - collection
      operationId: get_collection_price_history
      parameters:
      - name: start_date
        in: query
        description: 'Start date (ISO 8601: YYYY-MM-DD, inclusive). Defaults to end_date minus 30 days'
        required: false
        schema:
          type: string
      - name: end_date
        in: query
        description: 'End date (ISO 8601: YYYY-MM-DD, inclusive). Defaults to today'
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Collection price history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PriceHistoryEntryResponse'
        '400':
          description: Invalid date range (start_date > end_date)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_auth: []
  /collection/stats:
    get:
      tags:
      - collection
      operationId: get_collection_stats
      responses:
        '200':
          description: Collection stats for the authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionStatsResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_auth: []
...
components:
  schemas:
...
    BinderInfoResponse:
      type: object
      required:
      - name
      - card_count
      properties:
        card_count:
          type: integer
          format: int64
          minimum: 0
        name:
          type: string
    CardImportLineErrorResponse:
      type: object
      required:
      - line
      - field
      - value
      properties:
        field:
          type: string
        line:
          type: integer
          minimum: 0
        value:
          type: string
    CardImportResponse:
      type: object
      required:
      - id
      - status
      - source_lines
      - total_lines
      - processed_lines
      - errors
      - error_count
      - created_at
      properties:
        created_at:
          type: string
          description: RFC 3339 timestamp
        error_code:
          type:
          - string
          - 'null'
          description: |-
            Stable code of a failed import's error (`no_valid_line`, `internal`, …), translated by
            the clients. `null` unless the import failed.
        error_count:
          type: integer
          format: int32
          minimum: 0
        error_message:
          type:
          - string
          - 'null'
          description: Technical message of a failed import, for diagnosis only — never shown to the user.
        errors:
          type: array
          items:
            $ref: '#/components/schemas/CardImportLineErrorResponse'
          description: Truncated to at most 100 entries; `error_count` carries the true total.
        finished_at:
          type:
          - string
          - 'null'
          description: RFC 3339 timestamp
        id:
          type: string
        processed_lines:
          type: integer
          format: int32
          minimum: 0
        source_lines:
          type: integer
          format: int32
          description: Data lines read from the CSV, before the "Tokens" set filter and deduplication.
          minimum: 0
        status:
          type: string
          description: '`pending`, `running`, `completed` or `failed`.'
        total_lines:
          type: integer
          format: int32
          description: Cards to write, after deduplication — the denominator for `processed_lines`.
          minimum: 0
    CardImportStartedResponse:
      type: object
      required:
      - id
      properties:
        id:
          type: string
...
    CollectionCardResponse:
      type: object
      required:
      - set_code
      - collector_number
      - language_code
      - foil
      - name
      - rarity_code
      - scryfall_id
      - reserved
      properties:
        collection_entry:
          $ref: '#/components/schemas/CollectionEntryResponse'
          description: Present only when the card is owned by the authenticated user.
        collector_number:
          type: string
        foil:
          type: boolean
        image_back_url:
          type:
          - string
          - 'null'
          description: Back image of a double-faced card, like `image_url`; `null` for a single-faced card.
        image_url:
          type:
          - string
          - 'null'
          description: |-
            Front image, relative to the frontend (e.g. `/card-images/FDN_1_EN.webp?v=gatherer`);
            `null` while the card's image is pending.
        language_code:
          type: string
        name:
          type: string
        owner_count:
          type:
          - integer
          - 'null'
          format: int64
          description: |-
            Number of distinct users actually offering this card to trade (search mode only) —
            after applying each owner's visibility, trade binder selection and rarity filters. A
            card no one offers anymore never appears in search results at all.
          minimum: 0
        price_guide:
          $ref: '#/components/schemas/PriceGuideResponse'
        rarity_code:
          type: string
        reserved:
          type: boolean
          description: |-
            `true` if this card is engaged in an `ONE_ACCEPTED` or `FULLY_ACCEPTED` trade of its
            owner (the authenticated user in collection mode, or the `player_username`-filtered
            owner in search mode). Always `false` in the unscoped multi-owner search, where no
            single owner can be attributed.
        scryfall_id:
          type: string
        set_code:
          type: string
    CollectionEntryResponse:
      type: object
      required:
      - quantity
      - purchase_price
      - added_at
      properties:
        added_at:
          type: string
          description: RFC 3339 timestamp
        purchase_price:
          type: integer
          format: int32
          minimum: 0
        quantity:
          type: integer
          format: int32
          minimum: 0
    CollectionStatsResponse:
      type: object
      required:
      - total_cards
      - unique_cards
      - sets
      - binders
      properties:
        binders:
          type: array
          items:
            $ref: '#/components/schemas/BinderInfoResponse'
        price_trend_max:
          type:
          - integer
          - 'null'
          format: int32
          minimum: 0
        price_trend_min:
          type:
          - integer
          - 'null'
          format: int32
          minimum: 0
        sets:
          type: array
          items:
            $ref: '#/components/schemas/SetInfoResponse'
        total_cards:
          type: integer
          format: int64
          minimum: 0
        unique_cards:
          type: integer
          format: int64
          minimum: 0
...
    ErrorResponse:
      type: object
      description: |-
        Body of every error response. `error` is a technical message, for diagnosis only; clients
        translate `code` into what they show the user (ADR 0018).
      required:
      - error
      - code
      properties:
        code:
          type: string
          description: Stable snake_case identifier of the error (e.g. `binder_export`, `internal`).
        error:
          type: string
          description: Technical message, in English — never shown to the user as is.
...
    PaginatedCollectionResponse:
      type: object
      required:
      - items
      - total
      - page
      - page_size
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CollectionCardResponse'
        page:
          type: integer
          format: int32
          minimum: 0
        page_size:
          type: integer
          format: int32
          minimum: 0
        total:
          type: integer
          format: int64
          minimum: 0
...
    PriceGuideResponse:
      type: object
      properties:
        avg:
          type:
          - integer
          - 'null'
          format: int32
          minimum: 0
        low:
          type:
          - integer
          - 'null'
          format: int32
          minimum: 0
        trend:
          type:
          - integer
          - 'null'
          format: int32
          minimum: 0
    PriceHistoryEntryResponse:
      type: object
      required:
      - date
      - low
      - trend
      - avg
      properties:
        avg:
          type: integer
          format: int64
        date:
          type: string
          description: ISO 8601 date string (YYYY-MM-DD)
        low:
          type: integer
          format: int64
        trend:
          type: integer
          format: int64
    RarityCodeParam:
      type: string
      enum:
      - C
      - U
      - R
      - M
      - S
...
    SetInfoResponse:
      type: object
      required:
      - code
      - name
      properties:
        code:
          type: string
        name:
          type: string
...
    SortByParam:
      type: string
      enum:
      - avg
      - trend
      - set_code
      - language_code
      - added_at
    SortDirParam:
      type: string
      enum:
      - asc
      - desc
...
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...